Reimplement section field as a behavior
