+++ setup.py
@@ -60,6 +60,7 @@ setup(name='collective.nitf',
           'plone.app.textfield',
           'plone.app.upgrade',
           'plone.app.vocabularies',
+          'plone.behavior',
           'plone.dexterity',
           'plone.directives.form',
           'plone.formwidget.autocomplete',
@@ -67,6 +68,7 @@ setup(name='collective.nitf',
           'plone.indexer',
           'plone.portlets',
           'plone.registry',
+          'plone.supermodel',
           'plone.uuid',
           'Products.CMFPlone >=4.3',
           'Products.CMFQuickInstallerTool',
