+++ src/collective/nitf/configure.zcml
@@ -12,6 +12,7 @@
   <include file="profiles.zcml" />
   <include file="vocabulary.zcml" />
 
+  <include package=".behaviors" />
   <include package=".browser" />
   <include package=".portlets" />
   <include package=".upgrades" />
