+++ src/collective/nitf/profiles/default/types/collective.nitf.content.xml
@@ -32,6 +32,7 @@
   <element value="plone.app.dexterity.behaviors.metadata.IDublinCore" />
   <element value="plone.app.relationfield.behavior.IRelatedItems" />
   <element value="plone.app.lockingbehavior.behaviors.ILocking" />
+  <element value="collective.nitf.behaviors.interfaces.ISection" />
  </property>
 
  <!-- View information -->
