+++ src/collective/nitf/upgrades/v2000/profile/types/collective.nitf.content.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <object name="collective.nitf.content" meta_type="Dexterity FTI">
   <property name="behaviors" purge="false">
-    <element value="plone.app.lockingbehavior.behaviors.ILocking" />
+    <element value="collective.nitf.behaviors.interfaces.ISection" />
   </property>
   <property name="view_methods" purge="False">
     <element value="nitf_galleria" remove="True" />
